worker_rlimit_core should be off_t

+++ src/core/ngx_cycle.h
@@ -78,7 +78,7 @@ typedef struct {
 
      ngx_int_t                rlimit_nofile;
      ngx_int_t                rlimit_sigpending;
-     size_t                   rlimit_core;
+     off_t                    rlimit_core;
 
      int                      priority;
